small module cleanup and fmt refactor
