+++ src/lib.rs
@@ -1,17 +1,1 @@
 
-
-
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        let result = 2 + 2;
-        assert_eq!(result, 4);
-    }
-
-    // #[test]
-    // fn test_c_max() {
-    //     assert_eq!()
-    // }
-}
