+++ src/main.rs
@@ -1,17 +1,14 @@
-// use heuristic_opt::schrage::{Job, JobSequence};
-//use heuristic_opt::
-// use heuristic_opt::schrage;
+mod schrage;
 
-
-pub use crate::schrage::{Job, JobSequence};
+use schrage::{jobs, schrage::schrage};
 
 fn main() {
-    let js = JobSequence {
-        job_sequence: vec![Job::new(0, 10, 5), Job::new(1, 10, 123)],
+    let js = jobs::JobSequence {
+        job_sequence: vec![jobs::Job::new(0, 10, 5), jobs::Job::new(1, 10, 123)],
     };
     // println!("{}", Job::new(0, 10, 5));
     println!("{}", js);
-    let out = schrage(js);
+    let out = schrage(&vec![jobs::Job::new(0, 10, 5), jobs::Job::new(1, 10, 123)]);
 }
 // (0, 10, 5)
 // (30, 10, 5)
