--- src/schrage.rs
@@ -1,1 +1,0 @@
-pub mod schrage;
